Update Buttonscript to EventScript, added trigger example, renamed materials
