+++ CMPM125 Final/Assets/Materials/TriggerPreview.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 129a5872a89de964f95973ead63ae361
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
