+++ CMPM125 Final/Assets/Prefabs/TriggerExample.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1428b61b017616b42bd67c0f647f2423
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
